distance calculation and sorting by distance

+++ app/core/src/main/java/com/teamg/tourdeshot/core/api/local/LocalDTO.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -20,6 +22,6 @@ public class LocalDTO {
 
     private Coordinates coordinates;
 
-    private double distance;
+    private BigDecimal distance;
 
 }
